Rename vendor to product
